refactory: tidy up more stuff
