+++ app/(auth)/login/page.tsx
@@ -4,7 +4,7 @@ export default function LoginPage() {
   return (
     <>
       <section className={"h-screen flex justify-center items-center"}>
-        <AuthForm name={"Login"} />
+        <AuthForm name="Login" />
       </section>
     </>
   );
