+++ app/(web)/user/setting/page.tsx
@@ -1,3 +1,3 @@
-export default function ProfilePage() {
+export default function SettingsPage() {
   return <div>This is User Setting Pages</div>;
 }
