+++ components/Form/Form.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import type { ComponentProps } from "react";
 import type {
   FieldValues,
