+++ components/Supabase/SupabaseListener.tsx
@@ -4,15 +4,15 @@ import { useRouter } from "next/navigation";
 import { useEffect } from "react";
 import { useSupabase } from "./SupabaseProvider";
 
+type SupabaseListenerProps = {
+  serverAccessToken?: string;
+};
+
 // this component handles refreshing server data when the user logs in or out
 // this method avoids the need to pass a session down to child components
 // in order to re-render when the user's session changes
 // #elegant!
-export function SupabaseListener({
-  serverAccessToken,
-}: {
-  serverAccessToken?: string;
-}) {
+export function SupabaseListener({ serverAccessToken }: SupabaseListenerProps) {
   const { supabase } = useSupabase();
   const router = useRouter();
