+++ data/Idea/Index/index.ts
@@ -1,0 +1,2 @@
+export * from "./ideas";
+export * from "./tags";
