+++ parts/Idea/Index/index.ts
@@ -1,0 +1,6 @@
+export * from "./Header";
+export * from "./IdeaCard";
+export * from "./IdeasList";
+export * from "./SearchInput";
+export * from "./TagItem";
+export * from "./TagList";
